requirements.txt
flask==1.1.1
pyyaml==5.3.1